Make block count short

+++ src/java/org/apache/cassandra/index/sasi/disk/TokenTree.java
@@ -66,7 +66,7 @@ public class TokenTree
         validateMagic();
 
         tokenCount = file.getLong();
-        blockCount = file.getLong();
+        blockCount = file.getShort();
         treeMinToken = file.getLong();
         treeMaxToken = file.getLong();
     }
